Operator Overloading
I forgot that to overload operators, you have to implement the `ops::` Traits, it becomes less painful.

In order to overload an operator, one should declare an overloading type such as `type vec3_scalar = vec3`, then do the implementation on the typedef insetad of the original, now you can implement the a single trait multiple times as long as the types used are different.