In Order
In order it is way slower, if unordered I got a 40-50% speedup, Ordered is 20% at most

I feel the issue comes from the threadpool, it is not efficient at actually running tasks as it locks when there is a blocked thread, which most of the time is locked.

There are a lot of lock waiting time to interact with the "writing" vector, then the write to file is also very slow.

VerySleepy profiler does not show information I can properly understand.

Flamegraph does not work on Windows (unless you do some backflips and pirouetes to install DTrace which I will not do).