Raw Output is 50% better
Not good enough, it should be at least 75% faster (4 true cores) and maximum 87.5% faster (just the tracing part). 

Most probably we are still hitting a ton of write locks on the image.

The next best solution, would be to separate the image into scanlines, then write accordingly, but thats is a bit much more effort than I want to actually do.