5-10x improvement based on amount of tasks
Sleep test shows that in theory a 1 to 1 ratio of improvement per core is possible, probably there is more issues with locks for writing.

Will try the sender receiver test to write on the image

Main thread will parse the received colors and when its finished receiving pixels, it goes out.