Added Channel for data sending
For some bad reason, RwLock is not sync safe
Have to change into Mutex, as the reads are actually pretty locking in this case as it is constantly getting new tasks, 

Channel sending and processing of data is not really better than writing on an atomic pixel image. But is easier to understand.

In future atomic images will be better in GPUs as those should be quite free, passing through a single chokepoint with way more pixels and when the operations are magnitudes faster.

Still, sender receiver approach for a controlling thread looks liek a better genera approach.