BVH is consistently 1.5x to 2x slower (:
Although avg iterations go from 367 to 50% of bvh, it seems either the aabb calculations are a lot slower or the non momeory contiguous bvh structure is a nightmare for the cache.

Or there are other issues I just can't get to.
Removing the Ground sphere which could have been provoking a lot of extra steps did not solve anything...