First Perlin Noise Implementation
The "tiled noise" is  more pleasant but more obviously  not random.

The tiled noise seemed to com from a noise with a lower sample count that is repeated in the same form, but instead of getting index from th epermutations, they come directly from the truncated x/y/z to index values.
Cool.