Trilinear Interpolation
Added linear interpolation but I can't understand how an implementation fo a bilinear interpolation would work

In my head, you get 4 data points (f64 values here;
 - Do a similar interpolation between index 0-1 of i and j, then multiply by one of the 4 values of value array
 - Accumulator gets the value

That does, as well as the liniar interpolation that is just not worth for this. I guess it just depends on how many dimensions you have in the operation.

I guess bi-tri-... refers to the amount of dimensions you have in the original data structure, 

Also tried understanding cubic, bicubic and tricubic interpolation. A big failure I got nothing except that the RTOW applies the 3rd hermit cubic factor to the fractional parts that sets the t values for the trilinear interpolation and that rounds it off (applying a cubic to oneself always rounds values in that specific factor curve (x^2*(3-2*x)) -> 3x^2 - 2x^3

There are a lot of math fundamentals I am missing...