Make CornellBox Progress Gif
And delete the debug images
Now for the volumes and the final scene, which will probably be a shitfest.

One thing I could try for parallelization, make just a number of images at 1 sample, then average them, each threads takes care of an image, that would use a fuck ton of memory, but might be better?