Final Scene more or less correct
Noise is incorrect, at least it is not absorbing color properly at some parts

There is a BIG memory issues, it is increasing per iteration which most probably comes from the recursive nature of the ray function -> make it non recursive next.

The Cube made of random spheres is not quite there.

Sphere mapping was still incorrect, because I did not accound for x/z mapping to the same fractional as Y, which also was incorrect. The range for acos is [-1,1] instead of [0,1] in rust.

That would solve mostly everything that is/was wrong.

Probably the slowdown at higher samples is because of constant writes to memory seems, even at 1000 samples at some point it was moving around 7GB, when the only supposedly allocated per sample were Arc::clone of the shared info, not the full vectors or shared data. makes little sense.