Added Tracy Spans
Understood  a bit how to use Tracy with tracing crate

It is very easy, but I can't get the actuall underlying system/std function timings to know what is the actual cost inside a span

Based on
https://www.joshmcguigan.com/blog/cost-of-indirection-rust/

https://github.com/fralken/ray-tracing-the-next-week/blob/master/src/bvh.rs

It seems most of my cost is in indirection of dynamic objects.

Solution to keep "dynamism" but make it seamless
fun take_enum_object() -> {
match {
 type=> type.hit
}

all return a bool which is what is needed.

Enums should be expectedly 3x faster